Rename pretty-histogram to pretty-histogram-series
Every series plugin package ends in -series.
