+++ .circleci/config.yml
@@ -32,7 +32,7 @@ aliases:
         - packages/lwc-plugin-dual-range-histogram-series/node_modules
         - packages/lwc-plugin-hlc-area-series/node_modules
         - packages/lwc-plugin-image-watermark/node_modules
-        - packages/lwc-plugin-pretty-histogram/node_modules
+        - packages/lwc-plugin-pretty-histogram-series/node_modules
         - packages/lwc-plugin-rounded-candles-series/node_modules
         - packages/lwc-plugin-stacked-area-series/node_modules
         - packages/lwc-plugin-stacked-bars-series/node_modules
