+++ packages/lwc-plugin-pretty-histogram-series/.gitignore
