+++ packages/lwc-plugin-pretty-histogram-series/LICENSE
