+++ packages/lwc-plugin-pretty-histogram-series/NOTICE
