+++ packages/lwc-plugin-pretty-histogram-series/index.html
