+++ packages/lwc-plugin-pretty-histogram-series/src/data.ts
