+++ packages/lwc-plugin-pretty-histogram-series/src/example/index.html
