+++ packages/lwc-plugin-pretty-histogram-series/src/example/sample-data.ts
