+++ packages/lwc-plugin-pretty-histogram-series/src/options.ts
