+++ packages/lwc-plugin-pretty-histogram-series/src/pretty-histogram-series.ts
