+++ packages/lwc-plugin-pretty-histogram-series/src/renderer.ts
