+++ packages/lwc-plugin-pretty-histogram-series/src/vite-env.d.ts
