+++ packages/lwc-plugin-pretty-histogram-series/src/vite.config.js
