+++ packages/lwc-plugin-pretty-histogram-series/tsconfig.json
