+++ plugin-examples/src/index.html
@@ -30,7 +30,7 @@
             <li><a href="./plugins/grouped-bars-series/example/">GroupedBars Series</a></li>
             <li>Heatmap Series: <a href="./plugins/heatmap-series/example/">Example 1</a> / <a href="./plugins/heatmap-series/example/example2.html">Example 2</a></li>
             <li><a href="./plugins/hlc-area-series/example/">HLC Area Series</a></li>
-            <li><a href="./plugins/pretty-histogram/example/">Pretty Histogram</a></li>
+            <li><a href="./plugins/pretty-histogram-series/example/">Pretty Histogram</a></li>
             <li><a href="./plugins/lollipop-series/example/">Lollipop Series</a></li>
 			<li><a href="./plugins/rounded-candles-series/example/">Rounded Candle Series</a></li>
 			<li><a href="./plugins/background-shade-series/example/">Shaded Background Series</a></li>
