+++ plugin-examples/src/plugins/pretty-histogram-series/README.md
@@ -1,0 +1,7 @@
+# Pretty histogram series
+
+This plugin has graduated to a published package,
+[`@tradingview/lwc-plugin-pretty-histogram-series`](https://www.npmjs.com/package/@tradingview/lwc-plugin-pretty-histogram-series).
+Its source lives in [`packages/lwc-plugin-pretty-histogram-series`](../../../../packages/lwc-plugin-pretty-histogram-series).
+
+The `example/` folder here is the gallery page, which imports the package.
