+++ plugin-examples/src/plugins/pretty-histogram-series/example/index.html
