--- plugin-examples/src/plugins/pretty-histogram/README.md
@@ -1,7 +1,0 @@
-# Pretty histogram series
-
-This plugin has graduated to a published package,
-[`@tradingview/lwc-plugin-pretty-histogram`](https://www.npmjs.com/package/@tradingview/lwc-plugin-pretty-histogram).
-Its source lives in [`packages/lwc-plugin-pretty-histogram`](../../../../packages/lwc-plugin-pretty-histogram).
-
-The `example/` folder here is the gallery page, which imports the package.
